server/requirements.txt
appdata==2.1.2
asgiref==3.6.0
attrs==22.2.0
autopep8==2.0.1
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.0.3
colorama==0.4.6
colored==1.4.3
Django==4.1.7
django-cors-headers==3.14.0
django-environ==0.10.0
django-filter==22.1
django-json-field==0.5.7
django-jsonfield==1.4.1
django-queryable-properties==1.8.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
drf-spectacular==0.26.1
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
mypy-extensions==1.0.0
newsapi-python==0.2.7
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
pycodestyle==2.10.0
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
sqlparse==0.2.4
to-requirements.txt==1.3.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
